UPSTREAM: LSM: Split LSM preparation from initialization
Since we already have to do a pass through the LSMs to figure out if
exclusive LSMs should be disabled after the first one is seen as enabled,
this splits the logic up a bit more cleanly. Now we do a full "prepare"
pass through the LSMs (which also allows for later use by the blob-sharing
code), before starting the LSM initialization pass.

Change-Id: Ie7aeefd9dbbc954d9fd469adcdf4b2cee73a23dc
Signed-off-by: Kees Cook <keescook@chromium.org>
Signed-off-by: z3rokwq <z3rokwq@gmail.com>